+++ src/main/kotlin/com/example/kbocombo/crawler/combo/infra/NaverSportHandler.kt
@@ -1,18 +1,18 @@
-package com.example.kbocombo.crawler.infra
+package com.example.kbocombo.crawler.combo.infra
 
 import com.example.kbocombo.common.logInfo
-import com.example.kbocombo.crawler.application.NaverSportClient
-import com.example.kbocombo.crawler.utils.toTeamFilterCode
+import com.example.kbocombo.crawler.common.application.NaverSportClient
+import com.example.kbocombo.crawler.common.utils.toTeamFilterCode
 import com.example.kbocombo.game.domain.Game
 import com.example.kbocombo.game.infra.GameRepository
 import com.example.kbocombo.player.vo.Team
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Component
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
 
 @Component
 class NaverSportHandler(
