+++ src/main/kotlin/com/example/kbocombo/crawler/player/infra/KboPlayerDetailPageParser.kt
@@ -1,9 +1,9 @@
 package com.example.kbocombo.crawler.player.infra
 
 import com.example.kbocombo.common.logError
-import com.example.kbocombo.crawler.utils.toHittingHand
-import com.example.kbocombo.crawler.utils.toPitchingHand
-import com.example.kbocombo.crawler.utils.toPlayerDetailPosition
+import com.example.kbocombo.crawler.common.utils.toHittingHand
+import com.example.kbocombo.crawler.common.utils.toPitchingHand
+import com.example.kbocombo.crawler.common.utils.toPlayerDetailPosition
 import com.example.kbocombo.player.Player
 import com.example.kbocombo.player.vo.HittingHandType
 import com.example.kbocombo.player.vo.PitchingHandType
