+++ src/main/kotlin/com/example/kbocombo/crawler/player/infra/KboPlayerListPageCrawler.kt
@@ -1,6 +1,6 @@
 package com.example.kbocombo.crawler.player.infra
 
-import com.example.kbocombo.crawler.infra.KboHttpClient
+import com.example.kbocombo.crawler.common.application.KboHttpClient
 import com.example.kbocombo.player.vo.PlayerPosition
 import com.example.kbocombo.player.vo.Team
 import com.example.kbocombo.player.vo.WebId
